v.0.2.3 Update
Multiple Bug Fixes:
i.e. Recording was cut off, during breaks in the spoken sentences and not properly transcribed. For more information check out the "Long Recording" Section in README.md

UI Improvements:
Added Customization options.
Choose the Key-Binding, or the Audio Quality and which language will be used for the transcription.

Important:
The use of FFmpeg on your system is highly recommended!